feat: Add TaskDetailView component and update tasks
This commit adds a new component called TaskDetailView.vue and update tasks.

In TaskDetailView.vue, a new template is added with the TaskDetail component. The TaskDetail component is imported from "../components/TaskDetail.vue". The created hook emits a "message" event with a value of true.

In user.service.js, the listTaskById function is modified to remove the unused task parameter. The API call is changed from a POST request to a GET request. The Content-Type header is removed as it is not necessary for a GET request.

These changes are made to improve the functionality and user experience of the application.