feat: Add current due date for edit a task
This feat is possible view current date of due date when edit a task.
Not retrieving the date in the due date field is related to the format of the value being assigned to the editTask.duedate model. For the <input> element with type datetime-local, the value must be in the format YYYY-MM-DDTHH:MM.