feat: Add KanbanView component and kanban_module store
This commit adds a new component called KanbanView.vue and a new store module called kanban_module.js.

In KanbanView.vue, a new template is added with the KanbanComponent. The KanbanComponent is imported from "../components/Kanban.vue". The created hook emits a "message" event with a value of true.

In index.js, the kanban_module is imported and added to the modules object in the Vuex store.

These changes are made to introduce a Kanban view to the application and manage kanban-related data in the store.