test: Added Vitest
Vue.js is a progressive JavaScript framework for building user interfaces. It is designed to be incrementally adoptable, meaning is can use as much or as little of the framework as need. Vue.js is particularly known for its reactivity system, which efficiently tracks dependencies and updates the DOM.
When using Vitest with Vue.js, is can take advantage of Vitest's fast test runner and rich feature set to test Vue components.